RELEASE/requirements.txt
pycryptodome
pyotp
qrcode
Pillow
Pyzbar